Remove help flag from CLI
As the arguments need to be parsed before they can be interrogated,
the only way to see if the help flag has been added is to have all
the correct options first. And that's just silly. Now we can simply
try to parse the options and when that fails, we can show the help.